functions/careplan-predictor/requirements.txt
azure-functions
azure-cosmos
torch --index-url https://download.pytorch.org/whl/cpu
numpy
pandas